requirements/worker.txt
celery>=5
redis
qdrant-client
pypdf
openpyxl
python-docx
requests
